Fix showHidden, isRandom, and scope variable evaluation within _wvu_slots.html.
When evaluating tag attribute expressions with Dentaku, values that look like numbers are automatically converted to floating values, which means ‘1’ actually gets converted to 1.0, which if then converted to a string is ‘1.0’, which when converted to a boolean is FALSE. Radius::TagBinding#math_dictionary screws us. Confused yet? :)

We need to explicitly test for ` = ‘1.0’` when doing the fancy setup for the showHidden, isRandom, and scope variables.